Autenticação implementada

+++ src/main/java/br/com/michel/hercules/api/controller/StudentsController.java
@@ -3,7 +3,9 @@ package br.com.michel.hercules.api.controller;
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
+import javax.transaction.Transactional;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,11 +13,13 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestPart;
@@ -25,12 +29,16 @@ import org.springframework.web.multipart.MultipartFile;
 import br.com.michel.hercules.api.controller.dto.GradeDto;
 import br.com.michel.hercules.api.controller.dto.StudentBasicDto;
 import br.com.michel.hercules.api.controller.dto.StudentDto;
+import br.com.michel.hercules.api.controller.form.GradeForm;
 import br.com.michel.hercules.api.controller.form.StudentForm;
+import br.com.michel.hercules.api.controller.form.UpdateStudentForm;
+import br.com.michel.hercules.exceptions.InvalidRegisterException;
 import br.com.michel.hercules.model.Grade;
 import br.com.michel.hercules.model.Responsible;
 import br.com.michel.hercules.model.Student;
 import br.com.michel.hercules.repository.EmployeeRepository;
 import br.com.michel.hercules.repository.GradeRepository;
+import br.com.michel.hercules.repository.ProfileRepository;
 import br.com.michel.hercules.repository.ResponsibleRepository;
 import br.com.michel.hercules.repository.SchoolClassRepository;
 import br.com.michel.hercules.repository.StudentRepository;
@@ -55,6 +63,8 @@ public class StudentsController {
 	private SubjectRepository subjectRepository;
 	@Autowired
 	private GradeRepository gradeRepository;
+	@Autowired
+	private ProfileRepository profileRepository;
 	
 	@GetMapping("/students/{schoolClass}")
 	public List<StudentBasicDto> listSchoolClass(
@@ -81,7 +91,7 @@ public class StudentsController {
 	public StudentDto findStudent(
 			@PathVariable String register
 	) {
-		Student student = studentRepository.findByRegister(register);
+		Student student = studentRepository.findByRegister(register).get();
 		return new StudentDto(student);
 	}
 	
@@ -90,7 +100,13 @@ public class StudentsController {
 			@RequestBody GradeForm gradeForm,
 			@PathVariable String register
 	) {
-		Student student = studentRepository.findByRegister(register);
+		Optional<Student> optional = studentRepository.findByRegister(register);
+		
+		if(optional.isEmpty())
+			throw new InvalidRegisterException();
+		
+		Student student = optional.get();
+		
 		gradeForm.setStudent(student);
 		Grade grade = gradeForm.toGrade(employeeRepository, subjectRepository);
 		gradeRepository.save(grade);
@@ -98,14 +114,15 @@ public class StudentsController {
 		return ResponseEntity.created(null).body(new GradeDto(grade));
 	}
 	
-	@PostMapping(value = "/student", consumes = {"*/*"})
+	@PostMapping(value = "/student", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
 	public ResponseEntity<StudentDto> newStudent(
 			@RequestPart(name = "file", required = false) MultipartFile file,
 			@RequestPart("json") @Valid StudentForm studentForm
 	) {
 		studentForm.setPicture(file);
 		Student student = 
-				studentForm.toStudent(responsibleRepository, schoolClassRepository, resourcesPath);
+				studentForm.toStudent(responsibleRepository, schoolClassRepository,
+						profileRepository, resourcesPath);
 		studentRepository.save(student);
 		
 		URI uri = URI.create("/student/" + student.getRegister());
@@ -116,7 +133,13 @@ public class StudentsController {
 	public ResponseEntity<?> deleteStudent(
 			@PathVariable String register
 	) {
-		Student student = studentRepository.findByRegister(register);
+		Optional<Student> optional = studentRepository.findByRegister(register);
+		
+		if(optional.isEmpty())
+			throw new InvalidRegisterException();
+		
+		Student student = optional.get();
+		
 		Long responsibleId = student.getResponsible().getId();
 		studentRepository.delete(student);
 		
@@ -128,4 +151,21 @@ public class StudentsController {
 		return ResponseEntity.ok().build();
 	}
 	
+	@PutMapping("/student/{register}")
+	@Transactional
+	public ResponseEntity<StudentDto> updateStudent(
+			@RequestBody UpdateStudentForm form,
+			@PathVariable String register
+	) {
+		Optional<Student> optional = studentRepository.findByRegister(register);
+		
+		if(optional.isEmpty())
+			throw new InvalidRegisterException();
+		
+		Student student = optional.get();
+		
+		student = form.update(student, responsibleRepository, schoolClassRepository, studentRepository);
+		
+		return ResponseEntity.ok(new StudentDto(student));
+	}
 }
